+++ app/tag/controller.js
@@ -1,7 +1,18 @@
 const Tag = require('./model');
+const { policyFor } = require('../policy');
 
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can('create', 'Tag')) {
+      // <-- can create Tag
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk membuat tag`,
+      });
+    }
+
     let payload = req.body;
 
     let tag = new Tag(payload);
@@ -21,6 +32,15 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can('update', 'Tag')) {
+      // <-- can update Tag
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate tag`,
+      });
+    }
+
     let payload = req.body;
 
     let tag = await Tag.findOneAndUpdate({ _id: req.params.id }, payload, {
@@ -43,6 +63,15 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can('delete', 'Tag')) {
+      // <-- can delete Tag
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk menghapus tag`,
+      });
+    }
+
     let tag = await Tag.findOneAndDelete({ _id: req.params.id });
 
     return res.json(tag);
